# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kostum API",
        "description": "Документация",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/rebootpassword": {
            "post": {
                "tags": [
                    "Авторизация с api_token"
                ],
                "summary": "Поменять пароль",
                "description": "Поменять пароль(для авторизированных пользователей",
                "operationId": "rebootpassword",
                "parameters": [
                    {
                        "name": "api_token",
                        "in": "header",
                        "description": "Токен авторизации(api_token)",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Апи Токен",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "oldpassword, newpassword"
                                ],
                                "properties": {
                                    "oldpassword": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "123"
                                    },
                                    "newspassword": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "321"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Массив с: Message - Пароль был успешно заменён!, user - Данные юзера(новый апитокен)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Пароль был успешно изменен"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/auth/changePassword": {
            "post": {
                "tags": [
                    "Авторизация без api_token"
                ],
                "summary": "Помменять пароль",
                "description": "Помменять пароль",
                "operationId": "changePassword",
                "parameters": [
                    {
                        "name": "api_token",
                        "in": "header",
                        "description": "Токен авторизации(api_token)",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Апи Токен",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password, email, phone, address"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "123"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "321"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "321"
                                    },
                                    "code": {
                                        "type": "string",
                                        "format": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Массив с: Message - Пароль был успешно заменён, user - Данные юзера(новый апитокен)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}